refactor: move JuselessDataladAOMICID1000VBM to separate module
