+++ junifer/configs/juseless/datagrabbers/__init__.py
@@ -1,8 +1,11 @@
 """Provide imports for datagrabbers sub-package."""
 
 # Authors: Federico Raimondo <f.raimondo@fz-juelich.de>
+#          Leonard Sasse <l.sasse@fz-juelich.de>
+#          Synchon Mandal <s.mandal@fz-juelich.de>
 # License: AGPL
 
 from .ukb_vbm import JuselessDataladUKBVBM
 from .camcan_vbm import JuselessDataladCamCANVBM
 from .ixi_vbm import JuselessDataladIXIVBM
+from .aomic_id1000_vbm import JuselessDataladAOMICID1000VBM
